long long variables

+++ analysis/lle/lle.cu
@@ -2,9 +2,9 @@
 #pragma warning(push)
 #pragma warning(disable:6385)
 
-__device__ void LLE(Computation* data, int variation, void(* finiteDifferenceScheme)(numb*, numb*, numb*))
+__device__ void LLE(Computation* data, uint64_t variation, void(* finiteDifferenceScheme)(numb*, numb*, numb*))
 {
-    int stepStart, variationStart = variation * CUDA_marshal.variationSize;
+    uint64_t stepStart, variationStart = variation * CUDA_marshal.variationSize;
     LOCAL_BUFFERS;
     LOAD_ATTRIBUTES(true);
     if (data->isHires) TRANSIENT_SKIP_NEW(finiteDifferenceScheme);
